third_party/blink/web_tests/platform/generic/fast/css3-text/css3-text-decoration/getComputedStyle/getComputedStyle-text-decoration-style-expected.txt

Test to make sure text-decoration-style is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Ancestor should not inherit 'dashed' value from parent (fallback to initial 'solid' value):
PASS e.style.textDecorationStyle is ""
PASS computedStyle.textDecorationStyle is "solid"


Parent should cointain 'dashed':
PASS e.style.textDecorationStyle is "dashed"
PASS computedStyle.textDecorationStyle is "dashed"


JavaScript setter tests for valid, initial, invalid and blank values:
PASS e.style.textDecorationStyle is ""

Valid value 'solid':
PASS e.style.textDecorationStyle is "solid"
PASS computedStyle.textDecorationStyle is "solid"


Valid value 'double':
PASS e.style.textDecorationStyle is "double"
PASS computedStyle.textDecorationStyle is "double"


Valid value 'dotted':
PASS e.style.textDecorationStyle is "dotted"
PASS computedStyle.textDecorationStyle is "dotted"


Valid value 'dashed':
PASS e.style.textDecorationStyle is "dashed"
PASS computedStyle.textDecorationStyle is "dashed"


Valid value 'wavy':
PASS e.style.textDecorationStyle is "wavy"
PASS computedStyle.textDecorationStyle is "wavy"


Initial value:
PASS e.style.textDecorationStyle is "initial"
PASS computedStyle.textDecorationStyle is "solid"


Invalid value (this property accepts one value at a time only):
PASS e.style.textDecorationStyle is "initial"
PASS computedStyle.textDecorationStyle is "solid"


Invalid value (ie. 'unknown'):
PASS e.style.textDecorationStyle is "initial"
PASS computedStyle.textDecorationStyle is "solid"


Empty value (resets the property):
PASS e.style.textDecorationStyle is ""
PASS computedStyle.textDecorationStyle is "solid"


PASS successfullyParsed is true

TEST COMPLETE
